+++ frontend/src/components/admin/tabs/group-tab.vue
@@ -1,11 +1,7 @@
 <template>
   <div>
     <header class="tabHeader">
-      <AdminCreateUpdateDialog
-        table="Group"
-        max-width="20em"
-        :inputs="AdminGroupCreateUpdateInputs"
-      />
+      <AdminCreateUpdateDialog table="Group" max-width="20em" :inputs="AdminGroupCreateUpdateInputs" />
     </header>
     <AdminTable item-title="name" :headers="headers" :items="groups">
       <template #no-data>
@@ -18,32 +14,16 @@
         <RelationChips :pks="item.userSet" :objs="users" title-key="username" />
       </template>
       <template #[`item.librarySet`]="{ item }">
-        <RelationChips
-          :pks="item.librarySet"
-          :objs="normalLibraries"
-          title-key="path"
-        />
+        <RelationChips :pks="item.librarySet" :objs="normalLibraries" title-key="path" />
       </template>
       <template #[`item.actions`]="{ item }">
-        <AdminCreateUpdateDialog
-          table="Group"
-          :old-row="item"
-          max-width="20em"
-          :inputs="AdminGroupCreateUpdateInputs"
-          size="small"
-          density="compact"
-        />
-        <AdminDeleteRowDialog
-          table="Group"
-          :pk="item.pk"
-          :name="item.name"
-          size="small"
-          density="compact"
-        />
+        <AdminCreateUpdateDialog table="Group" :old-row="item" max-width="20em" :inputs="AdminGroupCreateUpdateInputs"
+          size="small" density="compact" />
+        <AdminDeleteRowDialog table="Group" :pk="item.pk" :name="item.name" size="small" density="compact" />
       </template>
     </AdminTable>
     <div id="groupHelp">
-      <h3>Group Logic</h3>
+      <h3>Group Access Logic</h3>
       <p>
         A library in no groups is accessible to every user and non-users if
         those are enabled.
@@ -79,11 +59,7 @@
           </tr>
           <tr>
             <td>
-              <GroupChip
-                :item="{ name: 'Include', exclude: false }"
-                title-key="name"
-                group-type
-              />
+              <GroupChip :item="{ name: 'Include', exclude: false }" title-key="name" group-type />
             </td>
             <td class="hidden">Hidden</td>
             <td class="hidden">Hidden</td>
@@ -91,11 +67,7 @@
           </tr>
           <tr>
             <td>
-              <GroupChip
-                :item="{ name: 'Exclude', exclude: true }"
-                title-key="name"
-                group-type
-              />
+              <GroupChip :item="{ name: 'Exclude', exclude: true }" title-key="name" group-type />
             </td>
             <td class="hidden">Hidden</td>
             <td class="see">Can See</td>
@@ -174,17 +146,21 @@ export default {
   margin-bottom: 2em;
   color: rgb(var(--v-theme-textSecondary));
 }
+
 #groupTable {
   border: solid thin;
   margin-top: 1em;
 }
+
 #groupTable th,
 #groupTable td {
   padding: 0.25em;
 }
+
 .see {
   background-color: rgb(var(--v-theme-includeGroup));
 }
+
 .hidden {
   background-color: rgb(var(--v-theme-excludeGroup));
 }